+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * New: Add support for drag and drop images into ksnip. ([#282](https://github.com/ksnip/ksnip/issues/282))
 * New: Insert embedded image into an existing screenshot. ([#293](https://github.com/ksnip/ksnip/issues/293))
 * New: Show screenshots in tabs. ([#298](https://github.com/ksnip/ksnip/issues/298))
+* New: Add Maximize Window Button in Print Preview. ([#190](https://github.com/ksnip/ksnip/issues/190))
 * Changed: Save As option was added and useInstantSave config was removed. ([#285](https://github.com/ksnip/ksnip/issues/285))
 * Fixed: Compilation error with Qt 5.15. ([#279](https://github.com/ksnip/ksnip/issues/279))
 * Fixed: Undo and redo translation reverts back to English. ([#209](https://github.com/ksnip/ksnip/issues/209))